--- a/execute/group_lookup.go
+++ b/execute/group_lookup.go
@@ -1,10 +1,8 @@
 package execute
 
 import (
-	"bytes"
 	"sort"
 
-	"github.com/apache/arrow/go/arrow"
 	"github.com/influxdata/flux"
 )
 
@@ -315,90 +313,44 @@ func (l *GroupLookup) Range(f func(key flux.GroupKey, value interface{})) {
 // for random access.
 type RandomAccessGroupLookup struct {
 	elements []*groupLookupElement
-	index    map[string]*groupLookupElement
+	index    map[uint64]*groupLookupElement
 }
 
 type groupLookupElement struct {
 	Key     flux.GroupKey
 	Value   interface{}
+	Next    *groupLookupElement
 	Deleted bool
 }
 
 // NewRandomAccessGroupLookup constructs a RandomAccessGroupLookup.
 func NewRandomAccessGroupLookup() *RandomAccessGroupLookup {
 	return &RandomAccessGroupLookup{
-		index: make(map[string]*groupLookupElement),
+		index: make(map[uint64]*groupLookupElement),
 	}
 }
 
-func (l *RandomAccessGroupLookup) idForKey(key flux.GroupKey) []byte {
-	var data [8]byte
+func (l *RandomAccessGroupLookup) idForKey(key flux.GroupKey) uint64 {
 	k, ok := key.(*groupKey)
 	if !ok {
 		k = newGroupKey(key.Cols(), key.Values())
 	}
-
-	var b bytes.Buffer
-	for _, i := range k.sorted {
-		c := k.cols[i]
-		vsz := 8
-		if k.values[i].IsNull() {
-			vsz = 1
-		} else if c.Type == flux.TBool {
-			vsz = 1
-		} else if c.Type == flux.TString {
-			vsz = len(k.values[i].Str())
-		}
-		// Extra 3 bytes for the two delimeters and the type.
-		b.Grow(len(c.Label) + vsz + 3)
-
-		b.WriteString(c.Label)
-		b.WriteByte(0)
-		b.WriteByte(byte(c.Type))
-
-		v := k.values[i]
-		if !v.IsNull() {
-			switch c.Type {
-			case flux.TInt:
-				arrow.Int64Traits.PutValue(data[:], v.Int())
-				b.Write(data[:arrow.Int64SizeBytes])
-			case flux.TUInt:
-				arrow.Uint64Traits.PutValue(data[:], v.UInt())
-				b.Write(data[:arrow.Uint64SizeBytes])
-			case flux.TFloat:
-				arrow.Float64Traits.PutValue(data[:], v.Float())
-				b.Write(data[:arrow.Float64SizeBytes])
-			case flux.TString:
-				b.WriteString(v.Str())
-			case flux.TBool:
-				if v.Bool() {
-					b.WriteByte(1)
-				} else {
-					b.WriteByte(0)
-				}
-			case flux.TTime:
-				arrow.Int64Traits.PutValue(data[:], int64(v.Time()))
-				b.Write(data[:arrow.Int64SizeBytes])
-			}
-		} else {
-			// Write an invalid byte if there is a null value
-			// so that we differentiate between an empty string
-			// and a null value.
-			b.WriteByte(^byte(0))
-		}
-		b.WriteByte(0)
-	}
-	return b.Bytes()
+	return k.hash64()
 }
 
 // Lookup will retrieve the value associated with the given key if it exists.
 func (l *RandomAccessGroupLookup) Lookup(key flux.GroupKey) (interface{}, bool) {
 	id := l.idForKey(key)
-	e, ok := l.index[string(id)]
-	if !ok || e.Deleted {
+	e, ok := l.index[id]
+	if !ok {
 		return nil, false
 	}
-	return e.Value, true
+	for ; e != nil; e = e.Next {
+		if !e.Deleted && key.Equal(e.Key) {
+			return e.Value, true
+		}
+	}
+	return nil, false
 }
 
 // LookupOrCreate will retrieve the value associated with the given key or,
@@ -416,13 +368,30 @@ func (l *RandomAccessGroupLookup) LookupOrCreate(key flux.GroupKey, fn func() in
 // Set will set the value for the given key. It will overwrite an existing value.
 func (l *RandomAccessGroupLookup) Set(key flux.GroupKey, value interface{}) {
 	id := l.idForKey(key)
-	e, ok := l.index[string(id)]
+	e, ok := l.index[id]
 	if !ok {
 		e = &groupLookupElement{
 			Key: key,
 		}
-		l.index[string(id)] = e
+		l.index[id] = e
 		l.elements = append(l.elements, e)
+	} else if !key.Equal(e.Key) {
+		// The present entry doesn't match the group key.
+		// This indicates a hash conflict so try to find
+		// the group key or add it.
+		for ; e.Next != nil; e = e.Next {
+			if key.Equal(e.Next.Key) {
+				break
+			}
+		}
+
+		if e.Next == nil {
+			e.Next = &groupLookupElement{
+				Key: key,
+			}
+			l.elements = append(l.elements, e.Next)
+		}
+		e = e.Next
 	}
 	e.Value = value
 	e.Deleted = false
@@ -431,7 +400,7 @@ func (l *RandomAccessGroupLookup) Set(key flux.GroupKey, value interface{}) {
 // Clear will clear the group lookup and reset it to contain nothing.
 func (l *RandomAccessGroupLookup) Clear() {
 	l.elements = nil
-	l.index = make(map[string]*groupLookupElement)
+	l.index = make(map[uint64]*groupLookupElement)
 }
 
 // Delete will remove the key from this GroupLookup. It will return the same
@@ -442,12 +411,17 @@ func (l *RandomAccessGroupLookup) Delete(key flux.GroupKey) (v interface{}, foun
 	}
 
 	id := l.idForKey(key)
-	e, ok := l.index[string(id)]
-	if !ok || e.Deleted {
+	e, ok := l.index[id]
+	if !ok {
 		return nil, false
 	}
-	e.Deleted = true
-	return e.Value, true
+	for ; e != nil; e = e.Next {
+		if !e.Deleted && key.Equal(e.Key) {
+			e.Deleted = true
+			return e.Value, true
+		}
+	}
+	return nil, false
 }
 
 // Range will iterate over all groups keys in a stable ordering.