--- a/go.mod
+++ b/go.mod
@@ -15,7 +15,7 @@ require (
 	github.com/benbjohnson/immutable v0.2.1
 	github.com/bonitoo-io/go-sql-bigquery v0.3.4-1.4.0
 	github.com/c-bata/go-prompt v0.2.2
-	github.com/cespare/xxhash v1.1.0
+	github.com/cespare/xxhash/v2 v2.1.1
 	github.com/dave/jennifer v1.2.0
 	github.com/denisenkom/go-mssqldb v0.0.0-20200428022330-06a60b6afbbc
 	github.com/eclipse/paho.mqtt.golang v1.2.0