--- a/stdlib/kafka/to.go
+++ b/stdlib/kafka/to.go
@@ -9,7 +9,7 @@ import (
 	"sort"
 	"time"
 
-	"github.com/cespare/xxhash"
+	"github.com/cespare/xxhash/v2"
 	"github.com/influxdata/flux"
 	"github.com/influxdata/flux/codes"
 	"github.com/influxdata/flux/execute"