This improves the performance of the random access group lookup by
utilizing xxhash to perform the hashing instead of relying on the hash
for a string value.

The primary benefit for this is that xxhash can compute a hash value
while maintaining a constant buffer and memory size. The previous
implementation utilized a string builder which could potentially result
in more allocations. This allows us to put an item into the Go map using
a uint64 key instead of relying on the string key.

It also is able to add that uint64 key to the group key itself with
minimal disruption. This helps with the common pattern of `Lookup`
followed by `Set` to prevent recomputing the hash twice.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written