@jsternberg 2021-01-20
Benchmark comparison:

```
benchmark                                                       old ns/op     new ns/op     delta
BenchmarkGroupLookup_LookupOrSet/Ordered-12                     33910         34407         +1.47%
BenchmarkGroupLookup_LookupOrSet/Interweaved-12                 57035         58410         +2.41%
BenchmarkGroupLookup_LookupOrSet/Unordered-12                   114957        118495        +3.08%
BenchmarkGroupLookup_LookupOrSet/Reversed-12                    101359        102998        +1.62%
BenchmarkGroupLookup_LookupOrSet/ManySplits-12                  101364        100998        -0.36%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Ordered-12         62217         16748         -73.08%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Interweaved-12     63066         16781         -73.39%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Unordered-12       63744         16757         -73.71%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Reversed-12        62708         17039         -72.83%
BenchmarkRandomAccessGroupLookup_LookupOrSet/ManySplits-12      62329         16776         -73.08%

benchmark                                                       old allocs     new allocs     delta
BenchmarkGroupLookup_LookupOrSet/Ordered-12                     11             11             +0.00%
BenchmarkGroupLookup_LookupOrSet/Interweaved-12                 49             49             +0.00%
BenchmarkGroupLookup_LookupOrSet/Unordered-12                   152            152            +0.00%
BenchmarkGroupLookup_LookupOrSet/Reversed-12                    160            160            +0.00%
BenchmarkGroupLookup_LookupOrSet/ManySplits-12                  129            129            +0.00%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Ordered-12         440            125            -71.59%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Interweaved-12     440            125            -71.59%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Unordered-12       440            125            -71.59%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Reversed-12        440            125            -71.59%
BenchmarkRandomAccessGroupLookup_LookupOrSet/ManySplits-12      440            125            -71.59%

benchmark                                                       old bytes     new bytes     delta
BenchmarkGroupLookup_LookupOrSet/Ordered-12                     10696         10696         +0.00%
BenchmarkGroupLookup_LookupOrSet/Interweaved-12                 9176          9176          +0.00%
BenchmarkGroupLookup_LookupOrSet/Unordered-12                   12728         12728         +0.00%
BenchmarkGroupLookup_LookupOrSet/Reversed-12                    11304         11304         +0.00%
BenchmarkGroupLookup_LookupOrSet/ManySplits-12                  13208         13208         +0.00%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Ordered-12         43594         17969         -58.78%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Interweaved-12     43598         17968         -58.79%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Unordered-12       43598         17972         -58.78%
BenchmarkRandomAccessGroupLookup_LookupOrSet/Reversed-12        43596         17969         -58.78%
BenchmarkRandomAccessGroupLookup_LookupOrSet/ManySplits-12      43597         17971         -58.78%
```

This makes the random access group lookup faster than the normal group lookup for all scenarios. Memory usage is still higher. Some potential follow ups:

1. Feature flag using random access as the default.
2. Investigate where the allocations come from to see if we can further reduce them. It is strange that the ordered insert for group lookup has 11 allocations and random access for the same one is still 125.